// nestjs-pong/src/interfaces/gameData.interface.ts
export interface gameData {
		ballx: number;
		bally: number;
		balldx: number;
		balldy: number;
		speed: number;
		goup1: boolean;
		goup2: boolean;
		godown1: boolean;
		godown2: boolean;
		posRack1: number;
		posRack2: number;
		position: number;
		radius: number;
		width: number;
		height: number;
		rackWidth: number;
		rackHeight: number;
		direction: number;
}

export class gameDataClass {
	public ballx: number;
	public bally: number;
	public balldx: number;
	public balldy: number;
	public speed: number;
	public goup1: boolean;
	public goup2: boolean;
	public godown1: boolean;
	public godown2: boolean;
	public posRack1: number;
	public posRack2: number;
	public position: number;
	public radius: number;
	public width: number;
	public height: number;
	public rackWidth: number;
	public rackHeight: number;
	public direction: number;
}